+++ packages/pigeon/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 1.0.20
+## 2.0.1
 
 * Adds support to Swift generation.
+
+## 2.0.0
+
+* Implements nullable parameters.
+* **BREAKING CHANGE** - Nonnull parameters to async methods on HostApis for ObjC
+  now have the proper nullability hints.
 
 ## 1.0.19